finish

+++ app/src/main/java/com/example/mycity/data/Place.kt
@@ -6,7 +6,7 @@ import androidx.annotation.StringRes
 data class Place(
     val id: Int,
     @StringRes val title: Int,
-    val rating: Double,
+    val rating: Int,
     @DrawableRes val imageRes: Int,
     val category: Categories,
     @StringRes val description: Int
